use s32 instead of long in DIRENTRY struct
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@23858 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ libpcsxcore/psxbios.c
@@ -193,10 +193,10 @@ typedef struct {
 
 /*
 typedef struct {
-	long next;
-	long func1;
-	long func2;
-	long pad;
+	s32 next;
+	s32 func1;
+	s32 func2;
+	s32 pad;
 } SysRPst;
 */
 
@@ -223,10 +223,10 @@ typedef struct {
 
 struct DIRENTRY {
 	char name[20];
-	long attr;
-	long size;
+	s32 attr;
+	s32 size;
 	struct DIRENTRY *next;
-	long head;
+	s32 head;
 	char system[4];
 };
